Improve LNC correction, adapt for CMI, estim alpha
This commit implements an improvement to the LNC algorithm:
Insted of using the orignal volume in the denominator of the correction
term, it uses the product of the projections on X and Y through PCA:
from V_rect/V to V_rect/V_rect_x * V_rect_y.

This modification is propagated to LNC adaptation for CMI and to alpha
estimation.